Handle escaped attribute values.
Addressing issue #1. The code already handled attributes with spaces in
them, but if used via the web (which is the expected use case) then
those spaces will be escaped. And it didn't know how to handle escaped
strings.

Implementing the decoding of the escaped string in Search Options, which
where all this query-string parsing is ending up. It needs a better
name.