Speed up multi-parameter searches
Fixes #3

With multi-parameter searches we were iterating through the full data
set for each parameter. This made no sense because the results have to
match **every** search parameter.

Now we reduce the data set each time we apply a search.

Doing this meant switching to `each` instead of `each_with_object`, so
I've changed the delegation in SearchParameters