Return no records if a query contains an invalid attribute
Fixes #7

Matchers now return 'false' to match? if the target object does not
respond to the method they try to call. This is a naive rescue, as I'm
rescuing everything, not just NoMethodCall. But I don't yet see a reason
to get fancier with my rescues.

This approach does introduce a problem, though. All future matchers have
to implement a way to handle rescues. Ruby inheritance won't check the
super-class for a rescue. So if I did

```ruby
class AbstractMatcher
  def match?
    false
  rescue
    false
  end
end
...

class ConcreteMatcher < AbstractMatcher
  def match?
    target.send(:non_existent_method)
  end
end
```

Then that would result in a NoMethodError exception. AbstractMatcher's
rescue exception never comes into play.

So there is almost certainly a better approach here, but I don't
currently know what it is. Hopefully I'm smarter in the future.

A side effect of this change is that I had to drop all the OpenStructs
in the matcher unit tests. OpenStruct doesn't raise NoMethodError when
you ask it about weird attributes. I've replaced it with a pickier
Struct.