Replace repeated rescues with Contingency Method
Davin pointed me to this section of the Ruby Style Guide

https://github.com/bbatsov/ruby-style-guide#contingency-methods

And that is certainly a better way to solve the problem I was trying to
solve in the previous commit.

If there's a chance that a concrete Matcher can blow up, it can call the
`match_with_contingency` method and pass it a block. If there's no
chance that the Matcher can blow up (see MatchAll) then it can just do
its own thing.