Fix boolean attribute searches
Fixes #10

MatchAttribute and MatchNoAttribute were focused on checking for nil,
which wasn't quite right.

MatchAttribute should match if the attribute returns anything. A return
value of `true` (the boolean value), "true" (the string) will both
match.  Anything that's not nil/false should match.

And MatchNoAttribute should match anything that Ruby considers falsey,
basically just nil and false.

This makes the matching code easier to write. For MatchAttribute it just
has to call the target's method. If it gets anything back, it matches.

MatchNoAttribute is the opposite. It calls the method and only returns
true if it gets back nil/false.