Rename and namespace all of the things.
Big commit here. It's like ripping the band-aid off.

When I extracted this code to a gem, I did a bad job. I missed several
things:

1. I didn't namespace any of the classes. This is bad.
2. I gave the gem a terrible name.
3. The API for clients was odd.

So, the gem is now renamed to QueryStringSearch, which isn't exactly
catchy but it's closer to what it actually does.

The namespacing change is the reason for most of the changes in this
commit. All classes in lib are now namespaced within QueryStringSearch.

And, finally, the method used by people using the gem is now

```ruby
QueryStringSearch.new(data_set, query_string).results
```

Which I think reads better.