Refactor 'reserved_words' to patterns, not strings
While doing some design for this feature:
https://github.com/umn-asr/query_string_search/issues/2 we found that
having the reserved_words be string was going to be a problem. This
commit is to refactor away those strings and replace them with regex
patterns.

These patterns will be more flexible going forward. The string approach
meant that the search value and the reserved word had to match exactly.
A regex approach will let us see if the search value matches a pattern.