Change parsing of attr/value pairs in query string
We were splitting on "=", which worked when we only cared about
equality. But to implement inequality operators we had to change this
code.

New approach uses a pattern that gets word, non-word, rest of string.
It'll be passed strings like:

title=balto
country=US|UK
enrolled>5
etc.

Test added to Search Options spec to make sure that it is parsing these
correctly.

I've also added a `operator` attribute to search_options. This stores
the "=", ">" or "<=", etc. I'm pretty sure this will be useful soon, but
I may be wrong.