Move configuration of matchers into MatcherFactory
Previous approach was to pass two values to the Matcher's initialize
function that it would know how to use. This was going to cause problems
as the # of things needed to configure a matcher grows.

Since we already have this nice factory, and factories are meant to
configure objects, I moved the configuration to the factory.