Rework Comparator/Comparison to use operators
In the previous version, Comparator didn't do anything but return a
Comparison, then the matchers used the Comparison methods directly.

In this commit, I'm separating matchers from Comparison and having them
interact with Comparator. Along with this, Comparator has a new method
to set the operator that Comparison will use. So we can do:

```
Comparator.using("=").does(1).compare_with?(1)
```

Exposing the ability to set the operator is how we're going to get
comparisons using inequality operators.

In order to keep the current matchers unchanged, I've turned `equal?`
and `contains?` into aliases that wrap around this new operator-focused
implementation.

The aliases also keep people from having to type ∈ (U+2208, by the way).
This is the technically correct operator to use here, as we are saying
that value 1 is a member of the set [1,2,3]. It's real math and
everything: https://en.wikipedia.org/wiki/Element_(mathematics)

But it's a pain to type, so people should continue to use the contains
alias.