Implement matching against collection attributes
In the context of ComparisonFactory, `config.other` contains the return
value of the data's method. So, in the case of our example, it contains
the return value of `movie.home_formats`.

We check `config.other` to see if it's an array. If it is, then we use
the SetComparison which checks to see if the data the user is looking
for is a sub-set of the source data.