Tighten up Equality comparison's criteria
Equality shouldn't just respond `true` to build_me. If the comparsions
get loaded in an order such that Equality is ahead of any of the other
comparisons, then Equality will get built incorrectly.

This fix is a bad fix, as it it requires Equality to know about the
build criteria of all the other comparisons. But it is a fix, so I'm
going to commit it.

Also, I noticed that the tests in this branch are 3x slower than the
tests in its parent branch. I think this is due to too much object
creation. That is something that needs to be fixed.