Gross fix to illustrate the problem
While implementing Comparisons I made the tests 3x as slow. This is
because every time I checked to see if a data row matched, I created a
new comparison via ComparisonFactory.

What should happen is that I create a comparison on the **first** time I
try to match data. That comparison can be memoized and then reused. For
each attempt to match I'll have to change the comparison's 'other'
property to the current data's value. The desired value and operator can
stay the same.

Just to prove that I was on the right track with this bugfix, I've
implemented a really nasty version of the code in this commit. This
brings the runtime of the tests back down to 1.1 seconds on my
machine, compared to 3.5+ seconds before the change.