Metaprogramming to creating 1 comparison per match
comparison is now an attribute on a matcher. When called the first time,
it creates an unfinshed `Comparison` object. It is unfinished because we
can not determine which comparison class is correct until `compare` is
called the first time. This is because ComparisonFactory needs an
example of the real data it will be matching against to determine what
comparison class to build.

The unfinished `Comparison` responds to `compare`, just like a real
comparison instance. But when it's called, it passes the data provided
on to ComparisonFactory so that a real comparison instance can be made.
Then, using the __setobj__ method of SimpleDelegator along with some
method re-definition the real comparison's `compare` method takes the
place of the unfinished Comparison's compare method.

This may be overly clever!

But, this cleverness has some advantages.

Tests are running in under 1.1 seconds.

Also, it lets the matcher do nothing more than call
`comparison.compare(actual_data)`, without having to check to see if
their comparison is 'real' or not.

It also totally removes Comparator from the library, since all it did
was configure Comparison instances. That is now done
automatically.

There are still some improvements to be made here, but with the
technique working I wanted to commit.