Always compare 2 values using set operations
After discussing the metaprogramming approach with @davinlagerroos, he
suggested a simpler approach. If we convert all tests for equality to
use the SetComparison, then the ComparisonFactory no longer needs to
know the data-set's return value for the attribute.

Old approach required the ComparisonFactory to know this. For the
examlpe of the query string "movies?q=home_formats=DVD", if
`movie.home_formats` returned an array, then ComparisonFactory would
build a SetComparison. But if `home_formats` returned a string, then
ComparisonFactory would return an EqualityComparison.

Under this new approach, ComparisonFactory will return a SetComparison
regardless. EqualityComparison no longer exists. The only other
comparison option is InequalityComparison, which will still be returned
if the operator is >, <, <= or >=.

This change led to a few more small tweaks. The `normalize` method in
AbstractComparison is now much simpler, and handles the Array conversion
that used to be in the SetComparison.