Support != operator in the Set comparison. Test.
Searches that use != will resolve to using Set comparisons. To do this
we need to tweak how the Set comparison works, allowing it to use one of
two boolean methods.

If the comparison uses the `:=` operator, then the result of the Set
intersection is checked to see if it contains `any?` records. The
presence of any records indicates that at least one match was made atd
the comparison was good.

If the comparison uses the `:!=` operator, then the result of the Set
intersection is checked to see if it contains `none?` records. A total
absence of results means no matches, so the comparison was good.

If added tests to cover this case.